+++ modules/test/playwright/fixtures/applicationsMenuPages.fixture.js
@@ -5,7 +5,7 @@
 
 import {test} from '@playwright/test';
 
-import {ApplicationsMenuPage} from '../pages/applicationsMenu.page';
+import {ApplicationsMenuPage} from '../pages/product-navigation-applications-menu/applicationsMenu.page';
 
 exports.test = test.extend({
 	_applicationsMenuPage: async ({page}, use) => {
