+++ modules/test/playwright/pages/object/objectDefinitions.page.js
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {ApplicationsMenuPage} from '../applicationsMenu.page';
+import {ApplicationsMenuPage} from '../product-navigation-applications-menu/applicationsMenu.page';
 
 export class ObjectDefinitionsPage {
 	constructor(page) {
@@ -46,7 +46,7 @@ export class ObjectDefinitionsPage {
 		await this.createObjectFolderButton.click();
 		const response = await responsePromise;
 
-		return await response.json();
+		return response.json();
 	}
 
 	async goto() {
