+++ modules/test/playwright/pages/product-navigation-applications-menu/applicationsMenu.page.js
