+++ modules/test/playwright/tests/global.setup.config.js
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export const globalSetupProjectConfig = {
+export const config = {
 	name: 'setup',
 	testMatch: 'tests/global.setup.js',
 	timeout: 60 * 1000,
